move code logic to its respective files
